Add /login/token to support Customer Login API
