+++ .changeset/mighty-candles-refuse.md
@@ -1,0 +1,5 @@
+---
+"@bigcommerce/catalyst-core": patch
+---
+
+Add /login/token endpoint to power Customer Login API
